backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
openai==1.78.1
python-dotenv==1.0.1
python-multipart==0.0.9
pypdf==4.2.0
python-docx==1.1.0
openpyxl==3.1.5
cryptography==43.0.3
psycopg[binary]==3.2.1
psycopg_pool
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
